set/altiben-ai-consultant/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.1.0
langchain-openai==0.0.2
pinecone-client==3.0.0
python-dotenv==1.0.0
pydantic==2.5.0
pypdf2==3.0.1